% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_ifrp.R
\name{AdaptiveIFRP}
\alias{AdaptiveIFRP}
\title{Compute adaptive intrinsic factor risk premia}
\usage{
AdaptiveIFRP(
  returns,
  factors,
  penalty_parameters,
  weighting_type = "c",
  relaxed = FALSE,
  check_arguments = TRUE
)
}
\arguments{
\item{returns}{(n_observations x n_returns)-dimensional matrix of test asset
excess returns.}

\item{factors}{(n_observations x n_factors)-dimensional matrix of factors.}

\item{penalty_parameters}{(n_parameters)-dimensional vector of penalty
parameter values from smallest to largest.}

\item{weighting_type}{character specifying the type of adaptive weights:
based on the correlation between factors and returns 'c'; based on the
regression coefficients of returns on factors 'b'; based on the first-step
intrinsic risk premia estimator 'a'; otherwise a vector of ones (any other
character). Default is 'c'.}

\item{relaxed}{boolean TRUE for re-fitting the model without shrinkage
post selection; FALSE otherwise. Default is FALSE.}

\item{check_arguments}{boolean TRUE if you want to check function arguments;
FALSE otherwise. Default is TRUE.}
}
\value{
(n_factors x n_parameters)-dimensional matrix of adaptive
intrinsic factor risk premia.
}
\description{
Computes adaptive intrinsic factor risk premia for various
penalty parameter values.
}
\examples{
# import package data on 15 risk factors and 42 test asset excess returns
factors = intrinsicFRP::factors[,-1]
returns = intrinsicFRP::returns[,-1]

weights = 1. / rowSums(stats::cor(factors, returns))
penalty_parameters = seq(0., 1., length.out = 100)

# compute adaptive intrinsic factor risk premia for each penalty parameter
aifrp = AdaptiveIFRP(
returns,
factors,
penalty_parameters
)

}
