% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factors}
\alias{factors}
\title{Factors - monthly observations from 01/1970 to 12/2021}
\format{
\subsection{\code{factors}}{

A data frame with 624 rows and 7 columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{Factor observations}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
factors
}
\description{
Monthly observations from 01/1970 to 12/2021 of
the Fama-French 5 factors the momentum factor.
}
\keyword{datasets}
