% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factors51}
\alias{factors51}
\title{yoo}
\format{
\subsection{\code{factors51}}{

A data frame with \code{519} rows and \code{51} columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{Factor observations}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
factors51
}
\description{
yoo
}
\keyword{datasets}
