% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FRPCpp}
\alias{FRPCpp}
\title{Compute factor risk premia}
\usage{
FRPCpp(
  returns,
  factors,
  misspecification_robust = TRUE,
  include_standard_errors = FALSE
)
}
\arguments{
\item{returns}{\verb{n_observations x n_returns}-dimensional matrix of test asset
excess returns.}

\item{factors}{\verb{n_observations x n_factors}-dimensional matrix of risk
factors.}

\item{misspecification_robust}{boolean \code{TRUE} for the
"misspecification-robust" Kar Robotti Shanken GLS approach using the
inverse covariance matrix of returns; \code{FALSE} for standard Fama-Mac-Beth
risk premia. Default is \code{TRUE}.}

\item{include_standard_errors}{boolean \code{TRUE} if you want to compute the
adaptive intrinsic factor risk premia HAC standard errors; \code{FALSE} otherwise.
Default is \code{FALSE}.}
}
\value{
a list containing \code{n_factors}-dimensional vector of factor
risk premia in \code{"risk_premia"}; if \code{include_standard_errors=TRUE}, then
it further includes \code{n_factors}-dimensional vector of factor risk
premia standard errors in \code{"standard_errors"}.
}
\description{
Computes Fama MacBeth (1973) factor risk premia or misspecification-robust factor
risk premia of Kan Robotti Shanken (2013) from data on test asset excess
returns and risk factors. Optionally computes the corresponding
heteroskedasticity and autocorrelation robust standard errors using the
Newey-West (1994) plug-in procedure to select the number of relevant lags,
i.e., \code{n_lags = 4 * (n_observations/100)^(2/9)}.
}
\examples{
factors = factors5FF[,-1]
returns = returnsMEBTM25[,-1]

frp = FactorRiskPremia(returns, factors, include_standard_errors = TRUE)

}
