% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns}
\alias{returns}
\title{Test Asset Excess Returns - monthly observations from 01/1970 to 12/2021}
\format{
\subsection{\code{returns}}{

A data frame with 624 rows and 26 columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{Asset excess returns}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
returns
}
\description{
Monthly excess returns on the 25 Size/Book-to-Market double sorted portfolios
from 01/1970 to 12/2021.
}
\keyword{datasets}
