% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returnsIND10}
\alias{returnsIND10}
\title{yoo}
\format{
\subsection{\code{returnsIND10}}{

A data frame with \code{624} rows and \code{11} columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{Factor observations}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
returnsIND10
}
\description{
yoo
}
\keyword{datasets}
