% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returnsMEBTM25}
\alias{returnsMEBTM25}
\title{yoo}
\format{
\subsection{\code{returnsMEBTM25}}{
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
returnsMEBTM25
}
\description{
yoo
}
\keyword{datasets}
