% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returnsOPINV25}
\alias{returnsOPINV25}
\title{yoo}
\format{
\subsection{\code{returnsOPINV25}}{

A data frame with \code{624} rows and \code{26} columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{Factor observations}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
returnsOPINV25
}
\description{
yoo
}
\keyword{datasets}
