% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{risk_free}
\alias{risk_free}
\title{Risk free - monthly observations from \code{07/1963} to \code{02/2024}}
\format{
\subsection{\code{risk_free}}{

A data frame with \code{624} rows and \code{2} columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{risk free observations}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
risk_free
}
\description{
Monthly observations from \code{07/1963} to \code{02/2024} of
the US risk free asset.
}
\keyword{datasets}
