% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_coefficients.R
\name{SDFCoefficients}
\alias{SDFCoefficients}
\title{SDF Coefficients}
\usage{
SDFCoefficients(
  returns,
  factors,
  include_standard_errors = FALSE,
  hac_prewhite = FALSE,
  check_arguments = TRUE
)
}
\arguments{
\item{returns}{A \verb{n_observations x n_returns}-dimensional matrix of test asset
excess returns.}

\item{factors}{A \verb{n_observations x n_factors}-dimensional matrix of factors.}

\item{include_standard_errors}{A boolean: \code{TRUE} if you want to compute the
SDF coefficients' HAC standard errors; \code{FALSE} otherwise.
Default is \code{FALSE}.}

\item{hac_prewhite}{A boolean indicating if the series needs pre-whitening by
fitting an AR(1) in the internal heteroskedasticity and autocorrelation
robust covariance (HAC) estimation. Default is \code{false}.}

\item{check_arguments}{A boolean: \code{TRUE} for internal check of all function
arguments; \code{FALSE} otherwise. Default is \code{TRUE}.}
}
\value{
A list containing \code{n_factors}-dimensional vector of SDF coefficients
in \code{"sdf_coefficients"}; if \code{include_standard_errors = TRUE}, then
it further includes \code{n_factors}-dimensional vector of SDF coefficients'
standard errors in \code{"standard_errors"};
}
\description{
Computes the misspecification-robust SDF coefficients of
Gospodinov-Kan-Robotti (2014) \url{https:#'doi.org/10.1093/rfs/hht135}:
\verb{GKRSDFcoefficients = (C' * V[R]^\{-1\} * C)^\{-1\} * C' * V[R]^\{-1\} * E[R]}
from data on factors \code{F} and test
asset excess returns \code{R}.
These notions of SDF coefficients minimize pricing errors:
\verb{argmin_\{d\} (E[R] - Cov[R,F] * d)' * V[R]^\{-1\} * (E[R] - Cov[R,F] * d)}.
Optionally computes the corresponding
heteroskedasticity and autocorrelation robust standard errors using the
Newey-West (1994) \url{doi:10.2307/2297912} plug-in procedure to select the
number of relevant lags, i.e., \code{n_lags = 4 * (n_observations/100)^(2/9)}.
}
\examples{
# import package data on 6 risk factors and 42 test asset excess returns
factors = intrinsicFRP::factors[,-1]
returns = intrinsicFRP::returns[,-1]

# compute GKR SDF coefficients and their standard errors
frp = SDFCoefficients(returns, factors, include_standard_errors = TRUE)

}
