% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{factors}
\alias{factors}
\title{Factors - monthly observations from \code{01/1970} to \code{12/2021}}
\format{
\subsection{\code{factors}}{

A data frame with \code{624} rows and \code{7} columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{Factor observations}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
factors
}
\description{
Monthly observations from \code{01/1970} to \code{12/2021} of
the Fama-French \code{5} factors the momentum factor.
}
\keyword{datasets}
