% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frp.R
\name{FRP}
\alias{FRP}
\title{Compute factor risk premia from data}
\usage{
FRP(
  returns,
  factors,
  krs = TRUE,
  include_standard_errors = FALSE,
  check_arguments = TRUE
)
}
\arguments{
\item{returns}{(n_observations x n_returns)-dimensional matrix of test asset
excess returns.}

\item{factors}{(n_observations x n_factors)-dimensional matrix of factors.}

\item{krs}{boolean true for Kar Robotti Shanken GLS approach using the inverse
covariance matrix of returns; false for standard Fama-Mac-Beth risk premia.}

\item{include_standard_errors}{boolean TRUE if you want to compute the
adaptive intrinsic factor risk premia HAC standard errors; FALSE otherwise.
Default is FALSE.}

\item{check_arguments}{boolean TRUE if you want to check function arguments;
FALSE otherwise. Default is TRUE.}
}
\value{
n_factors-dimensional vector of factor risk premia.
}
\description{
Computes Fama-MacBeth or misspecification-robust factor
risk premia of Kan Robotti Shanken (2013) from data on factors and test
asset excess returns. Optionally computes the corresponding
heteroskedasticity and autocorrelation robust standard errors using the
Newey-West (1994) plug-in procedure to select the number of relevant lags,
i.e., n_lags = 4 * (n_observations/100)^(2/9).
}
\examples{
# import package data on 15 risk factors and 42 test asset excess returns
factors = factors[,-1]
returns = returns[,-1]

# compute KRS factor risk premia and their standard errors
frp = FRP(returns, factors, include_standard_errors = TRUE)

}
