% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frp.R
\name{FRP}
\alias{FRP}
\title{Compute factor risk premia from data}
\usage{
FRP(
  returns,
  factors,
  krs = TRUE,
  include_standard_errors = FALSE,
  check_arguments = TRUE
)
}
\arguments{
\item{returns}{\verb{n_observations x n_returns}-dimensional matrix of test asset
excess returns.}

\item{factors}{\verb{n_observations x n_factors}-dimensional matrix of factors.}

\item{krs}{boolean \code{TRUE} for Kar Robotti Shanken GLS approach using the inverse
covariance matrix of returns; \code{FALSE} for standard Fama-Mac-Beth risk premia.
Default is \code{TRUE}.}

\item{include_standard_errors}{boolean \code{TRUE} if you want to compute the
adaptive intrinsic factor risk premia HAC standard errors; \code{FALSE} otherwise.
Default is \code{FALSE}.}

\item{check_arguments}{boolean \code{TRUE} if you want to check function arguments;
\code{FALSE} otherwise. Default is \code{TRUE}.}
}
\value{
a list containing \code{n_factors}-dimensional vector of factor
risk premia in \code{"risk_premia"}; if \code{include_standard_errors=TRUE}, then
it further includes \code{n_factors}-dimensional vector of factor risk
premia standard errors in \code{"standard_errors"}.
}
\description{
Computes Fama MachBeth (1973) or misspecification-robust factor
risk premia of Kan Robotti Shanken (2013) from data on factors and test
asset excess returns. Optionally computes the corresponding
heteroskedasticity and autocorrelation robust standard errors using the
Newey-West (1994) plug-in procedure to select the number of relevant lags,
i.e., \code{n_lags = 4 * (n_observations/100)^(2/9)}.
}
\examples{
# import package data on 15 risk factors and 42 test asset excess returns
factors = factors[,-1]
returns = returns[,-1]

# compute KRS factor risk premia and their standard errors
frp = FRP(returns, factors, include_standard_errors = TRUE)

}
