% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frp.R
\name{GiglioXiu2021RiskPremia}
\alias{GiglioXiu2021RiskPremia}
\title{Compute Factor Risk Premia using Giglio and Xiu (2021) Method}
\usage{
GiglioXiu2021RiskPremia(
  returns,
  factors,
  which_n_pca = 0,
  check_arguments = TRUE
)
}
\arguments{
\item{returns}{A \verb{n_observations x n_returns} matrix of asset returns.}

\item{factors}{A \verb{n_observations x n_factors} matrix of risk factors.}

\item{which_n_pca}{Method to determine the number of PCA components:
\code{0} for Giglio Xiu (2021) \url{doi:10.1086/714090},
\code{-1} for Ahn Horenstein (2013) \url{doi:10.3982/ECTA8968},
or any positive integer for a specific number of PCs.}

\item{check_arguments}{A boolean: \code{TRUE} for internal check of all function
arguments; \code{FALSE} otherwise. Default is \code{TRUE}.}
}
\value{
A list containing:
\describe{
\item{risk_premia}{A matrix of factor risk premia.}
\item{n_pca}{The number of principal components used.}
}
}
\description{
Computes the factor risk premia using the 3-step procedure of
Giglio and Xiu (2021) \url{doi:10.1086/714090}. The procedure consists in
\enumerate{
\item extracting p PCA from returns, where p is a consistent estimator of the
number of strong latent factors, 2) compute a cross-sectional regression of
average returns on the estimated betas of the p latent factors, and 3) compute
a time-series regression of observed factors on the p latent factors to obtain
their mimicking portfolio risk premia.
}
}
\examples{
\dontrun{
returns <- matrix(rnorm(200), nrow=20, ncol=10)
factors <- matrix(rnorm(60), nrow=20, ncol=3)
which_n_pca <- 0
result <- GiglioXiu2021RiskPremia(returns, factors, which_n_pca)
print(result)
}
}
