% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misspecification_test.R
\name{HJMisspecificationTest}
\alias{HJMisspecificationTest}
\title{Asset Prciing HJ model misspecification test}
\usage{
HJMisspecificationTest(
  returns,
  factors,
  hac_prewhite = FALSE,
  check_arguments = TRUE
)
}
\arguments{
\item{returns}{An \verb{n_observations x n_returns}-dimensional matrix of test asset
excess returns.}

\item{factors}{An \verb{n_observations x n_factors}-dimensional matrix of risk
factors.}

\item{hac_prewhite}{A boolean indicating if the series needs prewhitening by
fitting an AR(1) in the internal heteroskedasticity and autocorrelation
robust covariance (HAC) estimation. Default is \code{false}.}

\item{check_arguments}{A boolean \code{TRUE} if you want to check function arguments;
\code{FALSE} otherwise. Default is \code{TRUE}.}
}
\value{
A list containing the squared standardized HJ test statistic and
the corresponding p-value.
}
\description{
Computes the Hansen-Jagannathan (1997) \url{doi:10.1111/j.1540-6261.1997.tb04813.x}
model misspecification statistic and
p-value of an asset pricing model from test asset excess returns \code{R} and
risk factors \code{F}. The statistic is defined as:
\verb{HJDISTANCE = min_\{d\} (E[R] - beta * d)' * Var[R]^\{-1\} * (E[R] - beta * d)}
where \code{beta = Cov[R, F] * Var[F]^{-1}} are the regression coefficients of
test asset excess returns \code{R} on risk factors \code{F}.
Detailed computations and p-value calculations can be found in
Kan-Robotti (2008) <10.1016/j.jempfin.2008.03.003>.
}
\examples{
# Import package data on 6 risk factors and 42 test asset excess returns
factors = intrinsicFRP::factors[,-1]
returns = intrinsicFRP::returns[,-1]

# Compute the HJ model misspecification test
hj_test = HJMisspecificationTest(returns, factors)

}
