% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_frp.R
\name{OptimalAdaptiveFRP}
\alias{OptimalAdaptiveFRP}
\title{Compute optimal adaptive factor risk premia}
\usage{
OptimalAdaptiveFRP(
  returns,
  factors,
  penalty_parameters,
  weighting_type = "c",
  include_standard_errors = FALSE,
  plot_score = TRUE,
  check_arguments = TRUE
)
}
\arguments{
\item{returns}{\verb{n_observations x n_returns}-dimensional matrix of test asset
excess returns.}

\item{factors}{\verb{n_observations x n_factors}-dimensional matrix of factors.}

\item{penalty_parameters}{\code{n_parameters}-dimensional vector of penalty
parameter values from smallest to largest.}

\item{weighting_type}{character specifying the type of adaptive weights:
based on the correlation between factors and returns \code{'c'}; based on the
regression coefficients of returns on factors \code{'b'}; based on the first-step
intrinsic risk premia estimator \code{'a'}; otherwise a vector of ones (any other
character). Default is \code{'c'}.}

\item{include_standard_errors}{boolean \code{TRUE} if you want to compute the
adaptive intrinsic factor risk premia HAC standard errors; \code{FALSE} otherwise.
Default is \code{FALSE}.}

\item{plot_score}{boolean \code{TRUE} for plotting the model score; \code{FALSE} otherwise.
Default is \code{TRUE}.}

\item{check_arguments}{boolean \code{TRUE} if you want to check function arguments;
\code{FALSE} otherwise. Default is \code{TRUE}.}
}
\value{
a list containing the \code{n_factors}-dimensional vector of adaptive
factor risk premia in \code{"risk_premia"}; the optimal penalty
parameter value in \code{"penalty_parameter"}; the model score for each penalty
parameter value in \code{"model_score"};  if \code{include_standard_errors=TRUE}, then
it further includes \code{n_factors}-dimensional vector of factor risk
premia standard errors in \code{"standard_errors"}.
}
\description{
Computes optimal adaptive factor risk premia based on
pre-computed intrinsic factor risk premia for
various penalty parameter values. Tuning is performed via
minimizing the ratio between model misspecification and mis-identification.
Adaptive weights can be based on the correlation between
factors and returns, on the regression coefficients of returns on factors
or on the first-step intrinsic risk premia estimator. Optionally computes
the corresponding heteroskedasticity and autocorrelation robust standard
errors using the Newey-West (1994) plug-in procedure to select the number
of relevant lags, i.e., \code{n_lags = 4 * (n_observations/100)^(2/9)}.
}
\examples{
# import package data on 15 risk factors and 42 test asset excess returns
factors = intrinsicFRP::factors[,-1]
returns = intrinsicFRP::returns[,-1]

penalty_parameters = seq(0., 1., length.out = 100)

# compute optimal adaptive factor risk premia and their standard errors
aifrp = OptimalAdaptiveFRP(
returns,
factors,
penalty_parameters,
include_standard_errors = FALSE
)

}
