% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{returns}
\alias{returns}
\title{Test Asset Excess Returns - monthly observations from \code{01/1970} to \code{12/2021}}
\format{
\subsection{\code{returns}}{

A data frame with \code{624} rows and \code{43} columns:
\describe{
\item{Date}{Date in yyyymm format}
\item{...}{Asset excess returns}
}
}
}
\source{
\url{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}
}
\usage{
returns
}
\description{
Monthly excess returns on the \code{25} Size/Book-to-Market double sorted portfolios
and the \code{17} industry portfolios from \code{01/1970} to \code{12/2021}.
}
\keyword{datasets}
